pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["jonthus <joona.erkkila1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.6.0"
bio = "^1.3.3"
matplotlib = "^3.5.1"
pylint = "^2.12.2"
coverage = "^6.2"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.1.2"
pylint = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
